requirements.txt
python-dotenv
discord.py
yt-dlp
PyNaCl

